feat: add task runner
Define the Action trait which defines behavior for a generic Action to
be executed. No results are returned at this time.
Define a simple ClosureAction, which accepts a closure to run when a
rule is triggered.
Build a simple TaskRunner actor, which the PubSub actor uses to dispatch
tasks to be run.