feat: provide mvp task runner and rule set
Added a TaskRunner Actor that receives tasks to execute. These are
dispatched concurrently by the PubSub actor when rules are triggered.

To demonstrate this, I provide a ClosureAction that executes an
arbitrary closure when the rule is triggered.

I add 2 simple rules to the server: two print a message if a label is
added or removed from a PR.

This is a working MVP of github pilot.

Things to do next:
- Read rules from YAML (impl Deserilaize for Rule, basically)
- Add new types of Specialised actions for doing things on Github, incl
  adding labels, merging PRs, running actions etc.
- Specialised actions for mailing users
- Additional predicates

Then
- Track user stats / contributions?