feat: handle additional 20x HTTP response codes
Some github endpoints return 20x, which was returning an error. We now
deserialize the results correctly.

feat: collect error messages

When adding labels, we hit the API n times. If the label exists, we get
a 422, but we don't want to fail. We should keep on going.

So I collect the error messages (in case it's not a 'label exists'
error). But now how to know if the overall call passed or not?

Right now it indicates failure. I'm not sure how to fix this one beyond
hitting the endpoint with a "exists? delete, then add" logic