feat: add labels from yml files to repo (#18)
* feat: add label munging to cli

* Adds a new RepoProvider trait that provides an interface for all calls
  relevant to repo-wide HTTP requests.
* Add a `label` command to the CLI
* This has a few subcommands:
  * list - lists all labels
  * create - create a new label
  * delete - deletes a label
  * assign - give it a path to a yaml file containing label definitions,
    and it adds them as a batch command

* Add the HTTP requests to perform the commands mentioned above

Example

```
$ ghp-cli labels --help

Manipulate labels on a repo

USAGE:
    ghp-cli labels <SUBCOMMAND>

OPTIONS:
    -h, --help    Print help information

SUBCOMMANDS:
    assign    Assign labels to a repo
    create    Create a new label
    delete    Delete a label
    edit      Edit an existing label
    help      Print this message or the help of the given subcommand(s)
    list      List all labels on a repo
```

* fix: improve color repr in terminal tabels

feat: add different output formats

Adds ability to specify label outputs in text (default), YAML, ot JSON.

* feat: handle additional 20x HTTP response codes

Some github endpoints return 20x, which was returning an error. We now
deserialize the results correctly.

feat: collect error messages

When adding labels, we hit the API n times. If the label exists, we get
a 422, but we don't want to fail. We should keep on going.

So I collect the error messages (in case it's not a 'label exists'
error). But now how to know if the overall call passed or not?

Right now it indicates failure. I'm not sure how to fix this one beyond
hitting the endpoint with a "exists? delete, then add" logic