feat: add pr-merge to sdk and cli (#22)
* feat: add pr-merge to sdk and cli

This PR adds the ability to merge a PR from the gh-pilot cli. This obv
requires support form the API interface as well.

* feat: add cli aliases to merge options

Add long- and short- options for the merge arguments to the PR-merge
command.

* feat: handle bespoke error conditions for merge requests

Adds bespoke handling for the 422 HTTP response for merge that fail
validation.

* fix: only serialize non-none fields

The merge request fails if any MergeParameter values are "nil", so skip
serialization if they're None.