feat: action-merge
`merge_action` is a Github Pilot action that can automatically merge PRs
based on given criteria.

This is a very basic version of this action, and will not be as
full-featured as some other products like merge-queue, or bors;
but, using other actions, and labels as a messaging system, you _can_
build up some fairly complex behaviour.

I'll also add additional features to this action in future, after
allowing for a decent period to get feedback on how this action is
performing.

This action is the most complex written to date, so it has its own
module, rather than putting everything in one file.

=== Merge Action Parameters

The merge action behaviour is controlled with these parameters:

    acks_required: how many ACK-type messages are required,
    ack_patterns: A list of regex patterns to identify as ACK messages,
    reviews_required: how many approvals are needed,
    all_checks_must_pass: override Github's Run Check status by setting
this to false,
    merge_label: A label that determines the merge behaviour,
    perform_merge: When true, the action will execute the merge automatically IF the `merge_label` label is atached to the PR. When false, the action will ADD the merge label.

The builder has a sane set of defaults: 3 ACKS, 1 Review, all checks
must pass. The default merge label is `P-merge`, and it is added to the
PR by default, rather than auto-merging.

The default ACK regexes are `^(ut|t)?ACK$`, `^LGTM!?$`, `^:?\\+1:?$`, 👍

=== Executor

Adds a MergeAction executor type that holds references to trait
implementtors for fetching comments, contributors and pushing GH
actions. This makes it easy to test, although the **default**
implementation (needed for the Actix SystemService) will build a HTTP
client using envars for authentication credentials.

=== Other features

* Adds IssueId parser
* Adds tests for IssueId parser
* Implements the logic to pull the PR reviews from the provider and
  decide whether the PR can be merged based on # of approvals > min, and
  whether changes were requested.