feat: add check_suite event
adds deserialization support for the check_suite github event.
This insludes the App struct and app permissions which are inlded in the
event payload.

The Github docs appear to be out of date here, since there were several
permissions that were not included in the API docs -- I've added them to
the struct, but thi smay lead to deserialization bugs in future.