fix: check run pr objects
The list of pull requests referenced in the check_suite object are not
full PullRequest objects, so we needed to define a new struct to capture
this.

The Git references in that object are also slightly different, so also
added the GitReferenceShort struct.

The failing event now passes in a new regression test.