fix: extract related pr in merge action (#35)
Merge Action won't always be triggered on a PR event. But if we can, we
should still try and merge the PR if the conditions match. This means
extracting the related PR for the event.

This PR adds a `related_pr` method to GithubEvent which tries to dig out
the PR info from the different event types.