feat: add graphql query for org activity
The OrgActivity graph query seraches the given organisation for all
issue and PR events that were generated between two given dates.

A pagination object is also returned so that the full data set can be
returned.

The results can be used as a basis for generating a comprehensive
activity report